reoragnize some analysis scripts and fix plugin discovery
