+++ core/figure_tools/figure_tools/__init__.py
@@ -1,6 +1,7 @@
 from .anomaly_timeseries import AnomalyTimeSeries
 from .common_plots import observation_vs_model_maps, radiation_decomposition, \
-                          timeseries_and_anomalies, zonal_mean_vertical_and_column_integrated_map
+                          timeseries_and_anomalies, zonal_mean_vertical_and_column_integrated_map, \
+                          chuck_radiation
 from .figure import Figure
 from .global_mean_timeseries import GlobalMeanTimeSeries
 from .lon_lat_map import LonLatMap
